refactor(web): rename SnapshotAccordion.tsx to index.tsx to make it more consistent
