+++ veascan-web/src/components/SnapshotAccordion/index.tsx
